updated user model to include admins,updated categories and tags controller to not be able to delete if post are associated, created seed to seed database

+++ urban_connection/app/Http/Controllers/CategoriesController.php
@@ -93,6 +93,11 @@ class CategoriesController extends Controller
      */
     public function destroy( Category $category)
     {
+        if($category->posts->count()>0){
+            session()->flash('success', "Can't be deleted it has posts.");
+            return redirect()->back();
+
+        }
         $category->delete();
         session()->flash('success', 'Deleted successfully.');
         return redirect(route('categories.index'));
